requirements.txt
requests==2.25.1
python_telegram_bot==13.1
python-dotenv==0.15.0
